🗃Refactor: 모듈 의존성 제거
1. 이벤트 모듈을 제거하고 마스터 컴포넌트에 이벤트 바인딩 로직을 추가한다.
2. Storage 모듈의 불필요한 메서드를 제거하고 기능을 독립시킨다.
3. 마스터 컴포넌트의 수행 단계를 변경하고 옵저버 패턴을 명확히 주입한다.
